pyproject.toml
[tool.poetry]
name = "yt2raindrop"
version = "0.1.0"
description = ""
authors = ["Jake <jmetzger@arizona.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.163.0"
requests = "^2.32.3"
more-itertools = "^10.6.0"
cytoolz = "^1.0.1"
mypy = "^1.15.0"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = ["cytoolz.*", "googleapiclient.discovery.*", "google_auth_oauthlib.flow.*"]
follow_untyped_imports = true
ignore_missing_imports = true
